refactor: updated usage of TitleWithQuote component to remove duplication
