+++ app/[lang]/work/page.tsx
@@ -1,7 +1,21 @@
-import React, { ReactElement } from 'react';
+import { getTranslations } from 'next-intl/server';
+import React from 'react';
 
-import LiatoshynskyWork from '~/components/liatoshynsky-work/LiatoshynskyWork';
+import TitleWithQuote from '~/shared/components/title-with-quote/TitleWithQuote';
 
-export default function Work(): ReactElement {
-  return <LiatoshynskyWork />;
+export default async function Work() {
+  const t = await getTranslations('liatoshynskyWork');
+
+  return (
+    <TitleWithQuote
+      title={t('title-with-quote.title')}
+      quoteText={t('title-with-quote.quoteText')}
+      sourceText={{
+        title: t('title-with-quote.sourceText.title'),
+        data: t('title-with-quote.sourceText.data'),
+        place: t('title-with-quote.sourceText.place')
+      }}
+      color="black"
+    />
+  );
 }
